// src/app/data/enums.ts
export enum JOB {
  MERCENARY = "mercenary",
  ASSIST = "assist",
  ACROBAT = "acrobat",
  MAGICIAN = "magician",
  KNIGHT = "knight",
  BLADE = "blade",
  JESTER = "jester",
  RANGER = "ranger",
  RINGMASTER = "ringmaster",
  BILLPOSTER = "billposter",
  PSYCHIKEEPER = "psykeeper",
  ELEMENTOR = "elementor",
}

export enum PARAMS {
  HP = "hp",
  TIME = "time",
  ATTRIBUTE = "attribute",
  ATTACK = "attack",
  DURATION = "duration",
}